See #3: Move config from operators to utils
