+++ lib/index.js
@@ -1,7 +1,8 @@
 const operators = require('./operators');
+const loadConfig = require('./utils/config.util');
 
 const run = (command, ...args) => {
-  const config = operators.loadConfig();
+  const config = loadConfig();
 
   switch (command) {
     case 'n':
