+++ lib/operators/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   convert: require('./convert'),
-  loadConfig: require('./config'),
   generateMarkdown: require('./generate'),
 };
