+++ lib/utils/config.util.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
-const files = require('../utils/files.util');
+const files = require('./files.util');
 
 const load = () => {
   const rootDir = files.findRoot();
@@ -25,4 +25,7 @@ const getConfigValue = (dir, userConfig, defaultValue) => {
   return value;
 };
 
-module.exports = load;
+// This makes it so we can mock `load`
+// in tests
+module.exports = (...args) => module.exports.loadConfig(...args);
+module.exports.loadConfig = load;
